Added a preliminary version of a webots supervisor controller to record a vehicle's trajectory. Right now it just prints the vehicle's position ot the console. It should:
1. Capable of being turned on and off

2. Print to file

The first feature probably require setting a flag in the controller and checking at runtime. It may also require the supervisor
itself to be listening to the Homeo program. Alternatively, the flag could be set in the vehicle robot (which is already
listening to commands from the Homeo program) and the supervisor could just check it.

The second feature just requires setting a filename somewhere and printing to that file. Care must be taken with
opening/closing the file in case of crashes.

 Changes to be committed:
	new file:   src/Webots/Homeo-experiments/controllers/supervisor-trajectory/supervisor-trajectory.py
	modified:   src/Webots/Homeo-experiments/khepera-braitenberg-1-1-HOMEO.wbt
	modified:   src/Webots/Homeo-experiments/worlds/.khepera-braitenberg-1-1-HOMEO.wbproj
	new file:   src/Webots/Homeo-experiments/worlds/.khepera-braitenberg-2-HOMEO.wbproj
	modified:   src/Webots/Homeo-experiments/worlds/khepera-braitenberg-1-1-HOMEO.wbt
	modified:   src/Webots/Homeo-experiments/worlds/khepera-braitenberg-2-HOMEO.wbt