Added a final distance function method in GA general interface that returns  the distance between robot and target at simulation's end.
FIXME: The method read the robot's final distance from the last of the most recent
trajectory log. It would be more robust to ask the C++ supervisor controlling the
simulation to return the robot's final position.

FIXME 2: The target is currently specified in the Homeo GA general interface.
It would be more robust to indicate the target within Webots itself, and then ask
the C++ supervisor to return the distance.

 Changes to be committed:
	modified:   Simulator/HomeoGenAlgGui.py