Trajectory data files are now named according to the ID of the Homoestat generating them
In genetic algorithm simulations, each individual (i.e. a homoestat) is also given an ID
consiting of the generation number and the individual number with the generation,
in the form "gen-ind" where gen and ind are zero-padded integers.
The trajectory data file generated by the trajectory supervisor incorporates
the id in the data file's filename for later analysis.

Changes to be committed:
	modified:   src/Simulator/HomeoGenAlgGui.py
	modified:   src/Webots/Homeo-experiments/controllers/server-cpp/server-cpp.cpp
	modified:   src/Webots/Homeo-experiments/controllers/server-cpp/sockethandler.cpp
	modified:   src/Webots/Homeo-experiments/controllers/server-supervisor-cpp/robotworker.cpp
	modified:   src/Webots/Homeo-experiments/controllers/server-supervisor-cpp/sockethandler.cpp
	modified:   src/Webots/Homeo-experiments/controllers/supervisorTrajectory/supervisorTrajectory.py
	modified:   src/Webots/Homeo-experiments/worlds/khepera-braitenberg-2-HOMEO.wbt