First rudimentary prototype of a built-in robotic simulator
Homeo will include a fast and very simple  robotic simulator built on
top of the 2D physics package pyBox2D and the visualization package pyglet (which
is an openGl interface for python).

The simulator is built around to main classes: the model class holding the world and all
its objects (KheperaSimulation) and the view class allowing the visualization
(KheperaSimulationVisualizer).

In the current status, the simulator builds a world that includes a khepera-like robot
and a static object representing a light source, and the visualizer correctly shows them.
The visualizer can also zoom in and out responding according  to mouse-wheel input.

Movement of the robot is still not working correctly.

Changes to be committed:
	new file:   HomeoKhepSim/KheperaSimulator.py